Introduce new typeChecker for query
